+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- `AbstractAction::createNotFoundException()` because it was a bit pointless 🤦‍♂️
+
 ## [3.0.0] - 2023-03-11
 
 ### Added
@@ -96,12 +102,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
-- Updated `HttpResponse::send()` to accept an `HttpRequest`.  This should have been done previously :facepalm:
+- Updated `HttpResponse::send()` to accept an `HttpRequest`.  This should have been done previously 🤦‍♂️
 
 ## [1.0.0] - 2022-10-22
 
 First stable release.
 
+[unreleased]: https://github.com/danbettles/marigold/compare/v3.0.0...HEAD
 [3.0.0]: https://github.com/danbettles/marigold/compare/v2.5.0...v3.0.0
 [2.5.0]: https://github.com/danbettles/marigold/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/danbettles/marigold/compare/v2.3.3...v2.4.0